refresh JSON example
