+++ examples/experimental/json-browser/config/templates.js
@@ -1,0 +1,13 @@
+export default {
+  'website/3D Heatmap (HexagonLayer)': require('../../json-examples/3d-heatmap.json'),
+  'get-started/US map (GeoJsonLayer)': require('../../json-examples/us-map.json'),
+  'get-started/Dot Text (Scatterplot/TextLayer)': require('../../json-examples/dot-text.json'),
+  'experimental/Graph (BezierLayer)': require('../../json-examples/bezier-graph.json'),
+  // WEBSITE EXAMPLES AS JSON PAYLOADS
+  'website/GeoJSON (GeoJsonLayer)': require('../../json-examples/geojson.json'),
+  'website/Line (LineLayer)': require('../../json-examples/line.json'),
+  'website/Scatterplot (ScatterplotLayer)': require('../../json-examples/scatterplot.json'),
+  'website/Screen Grid (ScreenGridLayer)': require('../../json-examples/screen-grid.json'),
+  'website/TagMap (TextLayer)': require('../../json-examples/tagmap.json'),
+  'website/Plot (PlotLayer)': require('../../json-examples/plot.json')
+};
