+++ examples/experimental/json-browser/package.json
@@ -8,13 +8,18 @@
     "build": "NODE_ENV=production webpack --env.prod=true"
   },
   "dependencies": {
+    "@deck.gl/core": "^7.0.0-alpha",
     "@deck.gl/json": "^7.0.0-alpha",
+    "@deck.gl/layers": "^7.0.0-alpha",
+    "@deck.gl/react": "^7.0.0-alpha",
+    "@luma.gl/constants": "^7.0.0-alpha",
     "brace": "^0.11.1",
-    "deck.gl": "^7.0.0-alpha",
-    "react": "~16.2.0",
+    "brace": "^0.11.1",
+    "react": "~16.7.0",
     "react-ace": "^6.1.4",
-    "react-dom": "~16.2.0",
-    "react-map-gl": "^3.3.4"
+    "react-dom": "~16.7.0",
+    "react-map-gl": "^4.0.0",
+    "react-virtualized-auto-sizer": "^1.0.2"
   },
   "devDependencies": {
     "buble": "^0.19.3",
