--- examples/experimental/json-common/templates.js
@@ -1,13 +1,0 @@
-export default {
-  'website/3D Heatmap (HexagonLayer)': require('./examples/3d-heatmap.json'),
-  'get-started/US map (GeoJsonLayer)': require('./examples/us-map.json'),
-  'get-started/Dot Text (Scatterplot/TextLayer)': require('./examples/dot-text.json'),
-  'experimental/Graph (BezierLayer)': require('./examples/bezier-graph.json'),
-  // WEBSITE EXAMPLES AS JSON PAYLOADS
-  'website/GeoJSON (GeoJsonLayer)': require('./examples/geojson.json'),
-  'website/Line (LineLayer)': require('./examples/line.json'),
-  'website/Scatterplot (ScatterplotLayer)': require('./examples/scatterplot.json'),
-  'website/Screen Grid (ScreenGridLayer)': require('./examples/screen-grid.json'),
-  'website/TagMap (TextLayer)': require('./examples/tagmap.json'),
-  'website/Plot (PlotLayer)': require('./examples/plot.json')
-};
