+++ examples/experimental/json-examples/3d-heatmap.json
