+++ examples/experimental/json-examples/bezier-graph.json
