+++ examples/experimental/json-examples/plot.json
