+++ examples/experimental/json-examples/scatterplot.json
