+++ test/apps/json-pure-js/index.html
