+++ test/apps/json-pure-js/package.json
