+++ test/apps/json-pure-js/webpack.config.js
